LDEV-6288 restore segmented CHMnullSupport name for pre-7.1 session deserialisation
